+++ addon/locales/en/forms/fd-navigation.js
@@ -1,0 +1,7 @@
+export default {
+  caption: 'Navigation',
+  'caption-sheet': 'Tree node: ',
+  search: 'Search',
+  'root-folder': 'Root folder',
+  cancel: 'Cancel',
+};