+++ addon/locales/ru/components/fd-folder-editing-panel.js
@@ -1,0 +1,6 @@
+export default {
+  caption: 'Заголовок',
+  'caption-placeholder': 'Введите заголовок',
+  'sub-form': 'Страница',
+  'sub-folder': 'Вложенная папка',
+};