+++ addon/objects/uml-primitives/fd-uml-class.js
@@ -191,6 +191,8 @@ export let BaseClass = joint.shapes.basic.Generic.define('flexberry.uml.BaseClas
   }
 });
 
+joint.util.setByPath(joint.shapes, 'flexberry.uml.BaseClass', BaseClass, '.');
+
 /**
   Defines the JointJS element, which represents the UML class in the diagram.
 
@@ -201,6 +203,8 @@ export let BaseClass = joint.shapes.basic.Generic.define('flexberry.uml.BaseClas
   @constructor
 */
 export let Class = BaseClass.define('flexberry.uml.Class', {});
+
+joint.util.setByPath(joint.shapes, 'flexberry.uml.BaseClass', BaseClass, '.');
 
 joint.shapes.flexberry.uml.ClassView = joint.shapes.flexberry.uml.PrimitiveElementView.extend({
   template: [
@@ -305,19 +309,11 @@ joint.shapes.flexberry.uml.ClassView = joint.shapes.flexberry.uml.PrimitiveEleme
     }.bind(this));
 
     this.$box.find('.class-stereotype-input').on('blur', function(evt) {
-      let stereotypeText = $(evt.target).val();
-      let stereotype = this.normalizeStereotype(stereotypeText);
-      let rows = stereotypeText.split(/[\n\r|\r|\n]/);
-      let $stereotypeInput = this.$box.find('.class-stereotype-input');
-      $stereotypeInput.val(stereotype);
-      $stereotypeInput.prop('rows', rows.length);
-      let objectModel = this.model.get('objectModel');
-      objectModel.set('stereotype', stereotype);
-      this.paper.trigger('updaterepobj', objectModel, 'stereotype', stereotype);
-      this.updateRectangles();
+      this.showNormalizedStereotypeOnInput($(evt.target));
     }.bind(this));
 
     this.updateInputValue();
+    this.showNormalizedStereotypeOnInput(this.$box.find('.class-stereotype-input'));
 
     // Update the box position whenever the underlying model changes.
     this.model.on('change', this.updateBox, this);
@@ -366,9 +362,14 @@ joint.shapes.flexberry.uml.ClassView = joint.shapes.flexberry.uml.PrimitiveEleme
     let newWidth = 0;
     rects.forEach(function(rect) {
       if (this.markup.includes('flexberry-uml-' + rect.type + '-rect') && rect.element.inputElements) {
-        let $buffer = rect.element.inputElements.find('.input-buffer');
         let rectHeight = 0;
         let inputs = rect.element.inputElements.find('.' + rect.type + '-input');
+        let inputsDiv = inputs[0].parentElement;
+        if (! inputsDiv.parentElement || ! inputsDiv.parentElement.className.includes('joint-paper')) {
+          let jointPaper = $('.joint-paper')[0];
+          jointPaper.appendChild(inputsDiv);
+        }
+        let $buffer = rect.element.inputElements.find('.input-buffer');
         inputs.each(function() {
           let $input = $(this);
           $buffer.css('font-weight', $input.css('font-weight'));
@@ -491,5 +492,14 @@ joint.shapes.flexberry.uml.ClassView = joint.shapes.flexberry.uml.PrimitiveEleme
     }
 
     return stereotype;
+  },
+
+  showNormalizedStereotypeOnInput(element) {
+    let stereotypeText = element.val();
+    let stereotype = this.normalizeStereotype(stereotypeText);
+    let rows = stereotypeText.split(/[\n\r|\r|\n]/);
+    element.val(stereotype);
+    element.prop('rows', rows.length);
+    this.updateRectangles();
   }
 });