+++ tests/unit/controllers/new-platform-flexberry-web-designer-storage-type-l-test.js
@@ -4,9 +4,11 @@ moduleFor('controller:new-platform-flexberry-web-designer-storage-type-l', 'Unit
   // Specify the other units that are required for this test.
   // needs: ['controller:foo']
   needs: [
+    'controller:advlimit-dialog',
     'service:user-settings',
     'service:objectlistview-events',
-    'controller:colsconfig-dialog'
+    'controller:colsconfig-dialog',
+    'service:adv-limit',
   ]
 });